+++ packages/orq-rc/src/orq_ai_sdk/models/deployments.py
@@ -212,7 +212,7 @@ class AssistantMessage(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)
@@ -612,7 +612,7 @@ class MessagesAssistantMessage(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)