+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -372,7 +372,7 @@ class DeploymentsModelParameters(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)