+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -79,5 +79,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T12:31:28.307Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-03T06:54:06.707Z")
     r"""The date and time the resource was last updated"""