+++ packages/orq-rc/RELEASES.md
@@ -1099,3 +1099,13 @@ Based on:
 - [python v3.3.0-rc.28] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.28] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.28 - packages/orq-rc
+
+## 2025-04-03 06:55:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.528.1 (2.565.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.29] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.29] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.29 - packages/orq-rc